Move google_drive to conform to module naming convention
