+++ bin/cron/teacher_applications_to_gdrive
@@ -3,7 +3,7 @@ require_relative '../../lib/cdo/only_one'
 exit unless only_one_running?(__FILE__)
 
 require_relative '../../dashboard/config/environment'
-require 'cdo/google_drive'
+require 'cdo/google/drive'
 require 'honeybadger/ruby'
 
 # Squash an unhelpful warning that's causing a Honeybadger error
